+++ wcd/src/c3po/Textw.h
@@ -19,15 +19,15 @@ along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
 */
-   extern textw textwNewSize(int size);
+   extern textw textwNewSize(size_t size);
    extern textw textwNew(textw string);
    extern c3po_bool eqTextw(textw a,
                       textw b);
-   extern int inNamesetw(textw name,
+   extern size_t inNamesetw(textw name,
                         namesetw set);
    extern intset matchNamesetw(textw name,
                               namesetw set);
-   extern int matchCountNamesetw(textw name,
+   extern size_t matchCountNamesetw(textw name,
                                 namesetw set);
    extern textw concatw(textw a,
                       textw b);
@@ -39,16 +39,16 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
                        textw c,
                        textw d);
    extern textw repeatOnBufferw(textw pattern,
-                              int amount,
-                              int bufferNr);
+                              size_t amount,
+                              size_t bufferNr);
    extern textw repeatw(textw pattern,
-                      int amount);
-   extern textw spacesOnBufferw(int amount,
-                              int bufferNr);
-   extern textw spacesw(int amount);
-   extern textw TabOnBufferw(int amount,
-                           int bufferNr);
-   extern textw Tabw(int amount);
+                      size_t amount);
+   extern textw spacesOnBufferw(size_t amount,
+                              size_t bufferNr);
+   extern textw spacesw(size_t amount);
+   extern textw TabOnBufferw(size_t amount,
+                           size_t bufferNr);
+   extern textw Tabw(size_t amount);
    extern textw intToStringw(int i);
    extern textw doubleToStringw(double d);
 #endif
