+++ wcd/src/wcd.h
@@ -115,7 +115,7 @@ typedef char wcd_char;
 #endif
 
 FILE *wcd_fopen(const char *filename, const char *m, int quiet);
-void finddirs(char *dir, int *offset, FILE *outfile, int *use_HOME, nameset exclude, int quiet);
+void finddirs(char *dir, size_t *offset, FILE *outfile, int *use_HOME, nameset exclude, int quiet);
 void read_treefile(char *filename, nameset bd, int silent);
 void rmDirFromList(char *string, nameset n);
 void writeList(char *filename, nameset n);
